+++ server/src/error.rs
@@ -182,6 +182,8 @@ pub enum ApiError {
     ChatRateLimited,
     #[from(ignore)]
     AdvanceChatReadCursorError(tokio_postgres::error::Error),
+    #[from(ignore)]
+    AdvanceChatDeliveryCursorError(tokio_postgres::error::Error),
 }
 impl std::error::Error for ApiError {}
 fn handle_pg_error(err: &tokio_postgres::error::Error, name: &str) -> HttpResponse {
@@ -452,6 +454,9 @@ impl ResponseError for ApiError {
                 .body("Too many chat messages - slow down and try again shortly"),
             ApiError::AdvanceChatReadCursorError(ref err) => {
                 handle_pg_error(err, "AdvanceChatReadCursorError")
+            }
+            ApiError::AdvanceChatDeliveryCursorError(ref err) => {
+                handle_pg_error(err, "AdvanceChatDeliveryCursorError")
             }
         }
     }